requirements.txt
lxml==4.1.1
Pillow==5.0.0
python-docx==0.8.6
